restore mmap
